+++ sklearn/tree/export.py
@@ -8,6 +8,7 @@ This module defines export functions for decision trees.
 #          Noel Dawe <noel@dawe.me>
 #          Satrajit Gosh <satrajit.ghosh@gmail.com>
 #          Trevor Stephens <trev.stephens@gmail.com>
+#          Li Li <aiki.nogard@gmail.com>
 # License: BSD 3 clause
 
 import numpy as np
@@ -172,7 +173,8 @@ def export_graphviz(decision_tree, out_file=SENTINEL, max_depth=None,
             if len(sorted_values) == 1:
                 alpha = 0
             else:
-                alpha = int(np.round(255 * (sorted_values[0] - sorted_values[1]) /
+                alpha = int(np.round(255 * (sorted_values[0] -
+                                            sorted_values[1]) /
                                            (1 - sorted_values[1]), 0))
         else:
             # Regression tree or multi-output
@@ -330,7 +332,8 @@ def export_graphviz(decision_tree, out_file=SENTINEL, max_depth=None,
                         # Find max and min impurities for multi-output
                         colors['bounds'] = (np.min(-tree.impurity),
                                             np.max(-tree.impurity))
-                    elif tree.n_classes[0] == 1 and len(np.unique(tree.value)) != 1:
+                    elif (tree.n_classes[0] == 1 and
+                          len(np.unique(tree.value)) != 1):
                         # Find max and min values in leaf nodes for regression
                         colors['bounds'] = (np.min(tree.value),
                                             np.max(tree.value))
@@ -398,6 +401,16 @@ def export_graphviz(decision_tree, out_file=SENTINEL, max_depth=None,
         if out_file is None:
             return_string = True
             out_file = six.StringIO()
+
+        # Check length of feature_names before getting into the tree node
+        # Raise error if length of feature_names does not match
+        # n_features_ in the decision_tree
+        if feature_names is not None:
+            if len(feature_names) != decision_tree.n_features_:
+                raise ValueError("Length of feature_names, %d "
+                                 "does not match number of features, %d"
+                                 % (len(feature_names),
+                                    decision_tree.n_features_))
 
         # The depth of each node for plotting with 'leaf' option
         ranks = {'leaves': []}